Add API reference to README
We took the API reference from the React Docs guide on 'Typechecking
with PropTypes' and added it to this README.

Once this lands we can remove it from those docs and link to these docs.

Why? So that we make it clear that this is a separate module that can be
used with other libraries as well, and not just a feature of React. Also
hopefully this will make things more clear.

We will also open a `good-first-bug` issue for someone to improve this
README and turn it into a more complete API reference, with examples
using PropTypes with other libraries.

Fixes https://github.com/facebook/react/issues/9467