Downgrade deprecation error to warning
**what is the change?:**
Inlines a version of 'warning' that uses `console.warn` instead of `console.error`, and then uses that for the deprecation warning.

**NOTE: I'm not sure that this will only warn in development, which is a problem. We don't seem to have a '__DEV__' env. variable set up here, so I'm not sure how to limit the warnings to development?**

**why make this change?:**
This is a particularly noisy warning that was firing in 3rd party libraries, which then causes it to show up in product code. Users then can't silence it easily without updating their library. This makes it slightly less painful, hopefully, but still painful enough to get folks to update their code.

**test plan:**
`yarn test` - updated some tests to check for this.

**issue:**
https://github.com/facebook/react/issues/9398